Swapped KordiTextField to TextFormField with proper styling in ChangePasswordPage
